refactor(editor): Extract logic into custom hooks and utility modules
- EditorPane.tsx: Move spell check, Vim mode, editor handle, and keybindings logic to custom hooks
- Create useEditorVim.ts: Manages Vim mode lifecycle with initialization and disposal
- Create useEditorSpellCheck.ts: Handles spell check setup with typo-js dictionary loading
- Create useEditorHandle.ts: Exposes imperative editor methods via ref (text insertion, selection, scrolling, table operations)
- Create useEditorKeybindings.ts: Registers custom keybindings for formatting and table navigation
- Create editorOptions.ts: Centralizes Monaco editor configuration
- Create countWords.ts: Extracts word counting logic into utility function
- Remove unused imports (useImperativeHandle, useEffect, formatting/table utilities from main component)
- Simplify auto-continue logic by consolidating exit and continue actions
- Improve JSDoc formatting with proper punctuation and parameter descriptions
